Remove WindowMoveHandlare pImpl pattern
Introduces runtime overhead and makes debugging unecessary more painful
